qtvcp -make linuxcnc errors available to STATUS and ACTION
linuxcnc errors are a pain. It's always a race situation because
whom ever reads the error erases it.
With probing we need to know it the probe errored so we need away
to push error messages around to widgets.

right now screen options catches errors and pushes them to STATUS.
Really STATUS should poll errors and send them, but we can't do that
if qtvcp is embeded into other screens, as the other screens expect
to process the errors....